Fixes for api signing and when tone is null
1) Translate page opens up on android 5 and android 4. We need to have
a chat about whether each translation request will for sure have a tone
or if it can be null sometimes
2) Removed the floating email icon that were showing up at the bottom
right hand corner
3) Fixed the thing where text was getting cut off from the dialog box.
Did the thing you recommended: Shortened up the possible options when
"Not Sure" is clicked
4) Removed colon from headings
5) Added a space after the commas between fluent languages (if there
are more than 1)
6) When on a second verification activity, clicking the back button
(not the one on the top left, but the one at the bottom of the android
screen) was taking you to the previous verification. Now I disallow
that -- clicking the back button goes back to home screen.